requirements.txt
fastapi
importlib-resources
numpy
onnxruntime
tokenizers
uvicorn[standard]
joblib
regex
pandas
pillow
matplotlib
jinja2
pypdf2
fpdf
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit
matplotlib
tk

